+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 |kitty| is a feature full, cross-platform, *fast*, GPU based terminal emulator.
 To update |kitty|, :doc:`follow the instructions <binary>`.
 
-0.15.2 [future]
+0.16.0 [future]
 --------------------
 
 - A new :doc:`marks` feature that allows highlighting and scrolling to arbitrary